conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
accelerate=0.18.0=pypi_0
altair=4.2.2=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.0=pypi_0
apex=0.1=pypi_0
astor=0.8.1=pypi_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.7.1=pypi_0
beautifultable=0.7.0=pypi_0
blas=1.0=mkl
blinker=1.5=pypi_0
blis=0.7.9=pypi_0
boto3=1.26.2=pypi_0
botocore=1.29.2=pypi_0
braceexpand=0.1.7=pypi_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.2.0=pypi_0
catalogue=2.0.8=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.1=py39h74dc2b5_0
cfgv=3.3.1=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cloudpickle=2.2.0=pypi_0
cog=0.6.1=pypi_0
contexttimer=0.3.3=pypi_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
cython=0.29.32=pypi_0
cytoolz=0.12.0=pypi_0
dask=2022.10.2=pypi_0
datasets=1.11.0=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.6=pypi_0
distlib=0.3.6=pypi_0
docker=3.6.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docutils=0.19=pypi_0
easydict=1.10=pypi_0
einops=0.6.0=pypi_0
en-core-web-sm=3.2.0=pypi_0
entrypoints=0.4=pypi_0
evalai=1.3.16=pypi_0
exceptiongroup=1.1.1=pypi_0
executing=1.2.0=pypi_0
fairscale=0.4.4=pypi_0
fastapi=0.97.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.10.7=pypi_0
flatbuffers=22.10.26=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2022.10.0=pypi_0
ftfy=6.1.1=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.14.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.59.0=pypi_0
grpcio=1.50.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.7.0=pypi_0
httptools=0.5.0=pypi_0
huggingface-hub=0.0.12=pypi_0
identify=2.5.21=pypi_0
idna=3.3=pypi_0
imageio=2.22.3=pypi_0
imgcat=0.5.0=pypi_0
importlib-metadata=6.0.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
ipython=8.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
json-lines=0.5.0=pypi_0
jsonlines=3.0.0=pypi_0
jsonschema=4.17.3=pypi_0
kaggle=1.5.13=pypi_0
keras=2.10.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=14.0.6=pypi_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lmdb=0.94=pypi_0
locket=1.0.0=pypi_0
lxml=4.6.2=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
msgpack=1.0.0=pypi_0
msgpack-numpy=0.4.4.3=pypi_0
multiprocess=0.70.14=pypi_0
murmurhash=1.0.9=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=2.8.8=pypi_0
nodeenv=1.7.0=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
olefile=0.46=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
opendatasets=0.1.22=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
opentelemetry-api=1.18.0=pypi_0
opentelemetry-exporter-otlp=1.18.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.18.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.18.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.18.0=pypi_0
opentelemetry-proto=1.18.0=pypi_0
opentelemetry-sdk=1.18.0=pypi_0
opentelemetry-semantic-conventions=0.39b0=pypi_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pypi_0
pandas=1.5.1=pypi_0
parso=0.8.3=pypi_0
partd=1.3.0=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.6.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=22.2.2=py39h06a4308_0
platformdirs=3.1.1=pypi_0
plotly=5.13.1=pypi_0
portalocker=2.7.0=pypi_0
pre-commit=3.2.0=pypi_0
preshed=3.0.8=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=10.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.14.0=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.13=haa1d7c7_2
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python-magic=0.4.27=pypi_0
python-slugify=8.0.1=pypi_0
pytorch=1.12.0=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytorch-transformers=1.1.0=pypi_0
pytz=2022.6=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=24.0.1=pypi_0
readline=8.2=h5eee18b_0
redis=4.5.5=pypi_0
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.3.2=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.17.21=pypi_0
ruamel-yaml-clib=0.2.6=pypi_0
s3transfer=0.6.0=pypi_0
sacremoses=0.0.47=pypi_0
salesforce-lavis=1.0.2=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.10.0=pypi_0
seaborn=0.12.1=pypi_0
semver=2.13.0=pypi_0
sentencepiece=0.1.97=pypi_0
sentry-sdk=1.5.6=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.5.0=py39h06a4308_0
shortuuid=1.0.8=pypi_0
six=1.16.0=pyhd3eb1b0_1
smart-open=5.2.1=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
spacy=3.2.3=pypi_0
spacy-legacy=3.0.10=pypi_0
spacy-loggers=1.0.3=pypi_0
sqlite=3.39.3=h5082296_0
srsly=2.4.5=pypi_0
stack-data=0.6.2=pypi_0
starlette=0.27.0=pypi_0
streamlit=1.20.0=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=1.8=pypi_0
tensorflow=2.10.0=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
tensorpack=0.9.4=pypi_0
termcolor=1.1.0=pypi_0
text-unidecode=1.3=pypi_0
thinc=8.0.17=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
timm=0.4.9=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchvision=0.13.0=py39_cu113
tornado=6.2=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.8.1=pypi_0
typer=0.4.2=pypi_0
typing-extensions=4.1.1=pypi_0
tzdata=2022.7=pypi_0
tzlocal=4.3=pypi_0
urllib3=1.26.8=pypi_0
uvicorn=0.22.0=pypi_0
uvloop=0.17.0=pypi_0
validators=0.12.6=pypi_0
virtualenv=20.21.0=pypi_0
wandb=0.12.10=pypi_0
wasabi=0.10.1=pypi_0
watchdog=3.0.0=pypi_0
watchfiles=0.19.0=pypi_0
wcwidth=0.2.6=pypi_0
webdataset=0.2.48=pypi_0
websocket-client=1.5.1=pypi_0
websockets=11.0.3=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xxhash=3.1.0=pypi_0
xz=5.2.6=h5eee18b_0
yaspin=2.1.0=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
